Animations and effects

+++ Assets/Scripts/Reward.cs
@@ -11,7 +11,7 @@ public class Reward : MonoBehaviour
     public Rewards RewardType => _rewardType;
     public GameObject Target => _target;
 
-    public int RewardValue { get; private set; }
+    public float RewardValue { get; private set; }
 
 
     private void Update()
@@ -19,7 +19,7 @@ public class Reward : MonoBehaviour
         transform.position = Vector3.MoveTowards(transform.position, _target.transform.position, _speed * Time.deltaTime);
     }
 
-    public void Initialize(GameObject target, Rewards rewardType, int rewardValue)
+    public void Initialize(GameObject target, Rewards rewardType, float rewardValue)
     {
         _target = target;
         _rewardType = rewardType;
